Start adding rocket status in my profile: - Move API call to a different module - Refactor pages/Rockets.js - Refactor redux/rockets/rocket.js
