+++ src/redux/rockets/rocket.js
@@ -1,31 +1,26 @@
 import { createAsyncThunk } from '@reduxjs/toolkit';
+import getRockets from '../../getters/getRockets';
 
 const FETCH_ROCKETS = ' FETCH_ROCKETS';
 const BOOK_ROCKET = 'BOOK_ROCKET';
-const initialState = [];
-const url = 'https://api.spacexdata.com/v3/rockets';
+const initialState = {
+  status: 'idle',
+  rockets: [],
+};
 
 export const fetchRockets = createAsyncThunk(
   FETCH_ROCKETS,
-  async (post, { dispatch }) => {
-    const response = await fetch(url);
-    const data = await response.json();
-    const rockets = data.map((item) => ({
-      id: item.rocket_id,
-      name: item.rocket_name,
-      description: item.description,
-      image: item.flickr_images[0],
-      reserved: false,
-    }));
-    dispatch({
+  async (post, thunkAPI) => {
+    const payload = await getRockets();
+    thunkAPI.dispatch({
       type: FETCH_ROCKETS,
-      payload: rockets,
+      payload,
     });
   },
 );
 
 export const bookRocket = (id) => (dispatch, getState) => {
-  const state = getState().rockets;
+  const state = getState().rockets.rockets;
   const newState = state.map((rocket) => {
     if (rocket.id !== id) return rocket;
     const reserved = !rocket.reserved;
@@ -39,9 +34,16 @@ export const bookRocket = (id) => (dispatch, getState) => {
 const rocketReducer = (state = initialState, action) => {
   switch (action.type) {
     case FETCH_ROCKETS:
-      return action.payload;
+      return {
+        ...state,
+        status: 'succeeded',
+        rockets: action.payload,
+      };
     case BOOK_ROCKET:
-      return action.payload;
+      return {
+        ...state,
+        rockets: action.payload,
+      };
     default:
       return state;
   }
